# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/projects": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a paginated list of all projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Projects"
                ],
                "summary": "Admin - Get All Projects",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pagination.PaginatedResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Projects"
                ],
                "summary": "Admin - Create Project",
                "parameters": [
                    {
                        "description": "Project Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.CreateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/projects.Project"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/projects/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Projects"
                ],
                "summary": "Admin - Update Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Project Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projects.UpdateProjectRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/projects.Project"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a project",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin - Projects"
                ],
                "summary": "Admin - Delete Project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/public/projects": {
            "get": {
                "description": "Retrieve a list of all projects",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public - Projects"
                ],
                "summary": "Public - Get All Projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/projects.Project"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "images.Image": {
            "type": "object",
            "properties": {
                "altText": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "entityID": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isPrimary": {
                    "type": "boolean"
                },
                "mimeType": {
                    "type": "string"
                },
                "orderIndex": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "images.ImageUploadResult": {
            "type": "object",
            "properties": {
                "file_name": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "mime_type": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                }
            }
        },
        "pagination.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "total_data": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        },
        "pagination.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/pagination.Meta"
                }
            }
        },
...
        "projects.CreateProjectRequest": {
            "type": "object",
            "properties": {
                "content_markdown": {
                    "type": "string"
                },
                "demo_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "experience_id": {
                    "description": "UUID or null",
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/images.ImageUploadResult"
                    }
                },
                "is_featured": {
                    "type": "boolean"
                },
                "repo_url": {
                    "type": "string"
                },
                "skill_ids": {
                    "description": "UUIDs",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "start_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "projects.Project": {
            "type": "object",
            "properties": {
                "contentMarkdown": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "demoURL": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "experienceID": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/images.Image"
                    }
                },
                "isFeatured": {
                    "type": "boolean"
                },
                "repoURL": {
                    "type": "string"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/skills.Skill"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "projects.UpdateProjectRequest": {
            "type": "object",
            "properties": {
                "content_markdown": {
                    "type": "string"
                },
                "demo_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end_date": {
                    "type": "string"
                },
                "experience_id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/images.ImageUploadResult"
                    }
                },
                "is_featured": {
                    "type": "boolean"
                },
                "repo_url": {
                    "type": "string"
                },
                "skill_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "start_date": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "skills.Skill": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "iconURL": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}